RSSGenerator/requirements.txt
feedgen
gitpython